requirements.txt
airium==0.2.6
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
cloudinary==1.44.0
dj-database-url==2.3.0
Django==5.1.4
django-cloudinary-storage==0.3.0
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
idna==3.10
lxml==5.3.0
packaging==24.2
pandoc==2.4
pillow==11.1.0
plumbum==1.9.0
ply==3.11
psycopg2==2.9.10
psycopg2-binary==2.9.10
PyJWT==2.10.1
pypandoc==1.15
pypandoc_binary==1.15
python-docx==1.1.2
python-dotenv==1.1.0
pywin32==310
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.4.0
whitenoise==6.9.0
